requirements.txt
azure-identity == 1.15.0
msgraph-sdk == 1.1.0
msal == 1.26.0